+++ example/test/widget_test.dart
@@ -9,10 +9,11 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:example_app/main.dart';
 
+
 void main() {
   testWidgets('Sandbox app smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(const MySandboxApp());
-    expect(find.text('Frosted UI Kit Playground'), findsOneWidget);
+    expect(find.text('Frosted UI Kit'), findsOneWidget);
   });
 }